Move column header definitions into SearchResultsTable
