+++ src/components/Table/SearchResultsTable.test.jsx
@@ -2,12 +2,6 @@ import React from 'react';
 import { render, screen } from '@testing-library/react';
 
 import SearchResultsTable from './SearchResultsTable';
-import { createHeader } from './utils';
-import MultiSelectCheckBoxFilter from './Filters/MultiSelectCheckBoxFilter';
-import SelectFilter from './Filters/SelectFilter';
-
-import { serviceMemberAgencyLabel } from 'utils/formatters';
-import { BRANCH_OPTIONS, MOVE_STATUS_LABELS, MOVE_STATUS_OPTIONS } from 'constants/queues';
 
 const mockTableData = [
   {
@@ -36,83 +30,9 @@ function mockQueries() {
   };
 }
 
-const columns = [
-  createHeader('Move code', 'locator', {
-    id: 'locator',
-    isFilterable: false,
-  }),
-  createHeader('DOD ID', 'dodID', {
-    id: 'dodID',
-    isFilterable: false,
-  }),
-  createHeader(
-    'Customer name',
-    (row) => {
-      return `${row.lastName}, ${row.firstName}`;
-    },
-    {
-      id: 'customerName',
-      isFilterable: false,
-    },
-  ),
-  createHeader(
-    'Status',
-    (row) => {
-      return MOVE_STATUS_LABELS[`${row.status}`];
-    },
-    {
-      id: 'status',
-      isFilterable: true,
-      // eslint-disable-next-line react/jsx-props-no-spreading
-      Filter: (props) => <MultiSelectCheckBoxFilter options={MOVE_STATUS_OPTIONS} {...props} />,
-    },
-  ),
-  createHeader(
-    'Origin ZIP',
-    (row) => {
-      return row.originDutyLocationPostalCode;
-    },
-    {
-      id: 'originPostalCode',
-      isFilterable: true,
-    },
-  ),
-  createHeader(
-    'Destination ZIP',
-    (row) => {
-      return row.destinationDutyLocationPostalCode;
-    },
-    {
-      id: 'destinationPostalCode',
-      isFilterable: true,
-    },
-  ),
-  createHeader(
-    'Branch',
-    (row) => {
-      return serviceMemberAgencyLabel(row.branch);
-    },
-    {
-      id: 'branch',
-      isFilterable: true,
-      Filter: (props) => (
-        // eslint-disable-next-line react/jsx-props-no-spreading
-        <SelectFilter options={BRANCH_OPTIONS} {...props} />
-      ),
-    },
-  ),
-  createHeader(
-    'Number of Shipments',
-    (row) => {
-      return Number(row.shipmentsCount);
-    },
-    { id: 'shipmentsCount', isFilterable: true },
-  ),
-];
-
 describe('SearchResultsTable', () => {
   it('renders', () => {
-    render(<SearchResultsTable handleClick={() => {}} title="Results" columns={columns} useQueries={mockQueries} />);
+    render(<SearchResultsTable handleClick={() => {}} title="Results" useQueries={mockQueries} />);
     const results = screen.queryByText('Results (1)');
     expect(results).toBeInTheDocument();
     const locator = screen.queryByText('P33YJB');
